prepare release 0.0.4
