+++ Test/Graph/Cluster/test_graph_morphism.py
@@ -1,7 +1,7 @@
 import unittest
 from synkit.IO.data_io import load_from_pickle
 from synkit.IO.chem_converter import rsmi_to_its
-from synkit.Graph.ITS._misc import get_rc
+from synkit.Graph.ITS.its_decompose import get_rc
 from synkit.Graph.Cluster.graph_morphism import graph_isomorphism, subgraph_isomorphism
