+++ synkit/Graph/hydrogen.py
@@ -125,3 +125,40 @@ def explicit_hydrogen(graph: nx.Graph) -> nx.Graph:
         )  # Connect the hydrogen to its parent atom
 
     return new_graph
+
+
+def expand_hydrogens(graph: nx.Graph) -> nx.Graph:
+    """
+    For each node in the graph that has an 'hcount' attribute greater than zero,
+    adds the specified number of hydrogen nodes and connects them with edges that
+    have specific attributes.
+
+    Parameters
+    - graph (nx.Graph): A graph representing a molecule with nodes that can
+    include 'element', 'hcount', 'charge', and 'atom_map' attributes.
+
+    Returns:
+    - nx.Graph: A new graph with hydrogen atoms expanded.
+    """
+    new_graph = graph.copy()  # Create a copy to modify and return
+    atom_map = (
+        max(data["atom_map"] for _, data in graph.nodes(data=True))
+        if graph.nodes
+        else 0
+    )
+
+    # Iterate through each node to process potential hydrogens
+    for node, data in graph.nodes(data=True):
+        hcount = data.get("hcount", 0)
+        if hcount > 0:
+            for _ in range(hcount):
+                atom_map += 1
+                hydrogen_node = {
+                    "element": "H",
+                    "charge": 0,
+                    "atom_map": atom_map,
+                }
+                new_graph.add_node(atom_map, **hydrogen_node)
+                new_graph.add_edge(node, atom_map, order=(1.0, 1.0), standard_order=0.0)
+
+    return new_graph
